requirements.txt
alembic==1.7.6
click==8.0.3
dnspython==2.2.0
Flask==2.0.3
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
python-dotenv==0.19.2
Werkzeug==2.0.3
WTForms==3.0.1

Flask-Script
psycopg2